+++ www/src/components/shell/onboarding/cloud/InstallCli.js
@@ -2,7 +2,9 @@ import { useContext } from 'react'
 import { P } from 'honorable'
 import { ArrowTopRightIcon, Button } from 'pluralsh-design-system'
 
-import { CreateShellContext, OnboardingCard, NavSection } from '../CloudShell'
+import CreateShellContext from '../../../../contexts/CreateShellContext'
+
+import { NavSection, OnboardingCard } from '../../CloudShell'
 
 export default function InstallCli() {
   const { previous, next } = useContext(CreateShellContext)
@@ -15,14 +17,15 @@ export default function InstallCli() {
           color="text-light"
           marginBottom="large"
         >
-          Our <strong>Quickstart (CLI)</strong> documentation will bring you through the flow to get started on your local machine. After you’re done, move on to the next step to experience the rest of what Plural has to offer.
+          Our <strong>Quickstart (CLI)</strong> documentation will bring you through the flow to get started on your local machine. After you're done, move on to the next step to experience the rest of what Plural has to offer.
         </P>
         <Button
+          as="a"
           width="100%"
           endIcon={<ArrowTopRightIcon size={24} />}
-          onClick={() => {
-            window.open('https://docs.plural.sh/quickstart/getting-started')
-          }}
+          href="https://docs.plural.sh/quickstart/getting-started"
+          target="_blank"
+          rel="noopener noreferrer"
         >
           Read the documentation
         </Button>
